scanner: move `lfs.revListShas` to `scanner.RevListShas`
