+++ lfs/scanner.go
@@ -3,7 +3,6 @@ package lfs
 import (
 	"bufio"
 	"bytes"
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -15,6 +14,7 @@ import (
 	"time"
 
 	"github.com/github/git-lfs/git"
+	"github.com/github/git-lfs/scanner"
 	"github.com/github/git-lfs/tools"
 	"github.com/rubyist/tracerx"
 )
@@ -66,46 +66,10 @@ type indexFile struct {
 
 var z40 = regexp.MustCompile(`\^?0{40}`)
 
-type ScanningMode int
-
-const (
-	ScanRefsMode         = ScanningMode(iota) // 0 - or default scan mode
-	ScanAllMode          = ScanningMode(iota)
-	ScanLeftToRemoteMode = ScanningMode(iota)
-)
-
-type ScanRefsOptions struct {
-	ScanMode         ScanningMode
-	RemoteName       string
-	SkipDeletedBlobs bool
-	nameMap          map[string]string
-	mutex            *sync.Mutex
-}
-
-func (o *ScanRefsOptions) GetName(sha string) (string, bool) {
-	o.mutex.Lock()
-	name, ok := o.nameMap[sha]
-	o.mutex.Unlock()
-	return name, ok
-}
-
-func (o *ScanRefsOptions) SetName(sha, name string) {
-	o.mutex.Lock()
-	o.nameMap[sha] = name
-	o.mutex.Unlock()
-}
-
-func NewScanRefsOptions() *ScanRefsOptions {
-	return &ScanRefsOptions{
-		nameMap: make(map[string]string, 0),
-		mutex:   &sync.Mutex{},
-	}
-}
-
 // ScanRefs takes a ref and returns a slice of WrappedPointer objects
 // for all Git LFS pointers it finds for that ref.
 // Reports unique oids once only, not multiple times if >1 file uses the same content
-func ScanRefs(refLeft, refRight string, opt *ScanRefsOptions) ([]*WrappedPointer, error) {
+func ScanRefs(refLeft, refRight string, opt *scanner.ScanRefsOptions) ([]*WrappedPointer, error) {
 	s, err := ScanRefsToChan(refLeft, refRight, opt)
 	if err != nil {
 		return nil, err
@@ -123,12 +87,12 @@ func ScanRefs(refLeft, refRight string, opt *ScanRefsOptions) ([]*WrappedPointer
 // ScanRefsToChan takes a ref and returns a channel of WrappedPointer objects
 // for all Git LFS pointers it finds for that ref.
 // Reports unique oids once only, not multiple times if >1 file uses the same content
-func ScanRefsToChan(refLeft, refRight string, opt *ScanRefsOptions) (*PointerChannelWrapper, error) {
+func ScanRefsToChan(refLeft, refRight string, opt *scanner.ScanRefsOptions) (*PointerChannelWrapper, error) {
 	if opt == nil {
-		opt = NewScanRefsOptions()
+		opt = scanner.NewScanRefsOptions()
 	}
 	if refLeft == "" {
-		opt.ScanMode = ScanAllMode
+		opt.ScanMode = scanner.ScanAllMode
 	}
 
 	start := time.Now()
@@ -136,12 +100,12 @@ func ScanRefsToChan(refLeft, refRight string, opt *ScanRefsOptions) (*PointerCha
 		tracerx.PerformanceSince("scan", start)
 	}()
 
-	revs, err := revListShas(refLeft, refRight, opt)
+	revs, errs, err := scanner.RevListShas(refLeft, refRight, opt)
 	if err != nil {
 		return nil, err
 	}
 
-	smallShas, err := catFileBatchCheck(revs)
+	smallShas, err := catFileBatchCheck(revs, errs)
 	if err != nil {
 		return nil, err
 	}
@@ -233,7 +197,6 @@ func ScanIndex(ref string) ([]*WrappedPointer, error) {
 
 	allRevsErr := make(chan error, 5) // can be multiple errors below
 	allRevsChan := make(chan string, 1)
-	allRevs := NewStringChannelWrapper(allRevsChan, allRevsErr)
 	go func() {
 		seenRevs := make(map[string]bool, 0)
 
@@ -262,7 +225,7 @@ func ScanIndex(ref string) ([]*WrappedPointer, error) {
 		close(allRevsErr)
 	}()
 
-	smallShas, err := catFileBatchCheck(allRevs)
+	smallShas, err := catFileBatchCheck(allRevsChan, allRevsErr)
 	if err != nil {
 		return nil, err
 	}
@@ -291,139 +254,6 @@ func ScanIndex(ref string) ([]*WrappedPointer, error) {
 
 	return pointers, err
 
-}
-
-// Get additional arguments needed to limit 'git rev-list' to just the changes
-// in revTo that are also not on remoteName.
-//
-// Returns a slice of string command arguments, and a slice of string git
-// commits to pass to `git rev-list` via STDIN.
-func revListArgsRefVsRemote(refTo, remoteName string) ([]string, []string) {
-	// We need to check that the locally cached versions of remote refs are still
-	// present on the remote before we use them as a 'from' point. If the
-	// server implements garbage collection and a remote branch had been deleted
-	// since we last did 'git fetch --prune', then the objects in that branch may
-	// have also been deleted on the server if unreferenced.
-	// If some refs are missing on the remote, use a more explicit diff
-
-	cachedRemoteRefs, _ := git.CachedRemoteRefs(remoteName)
-	actualRemoteRefs, _ := git.RemoteRefs(remoteName)
-
-	// Only check for missing refs on remote; if the ref is different it has moved
-	// forward probably, and if not and the ref has changed to a non-descendant
-	// (force push) then that will cause a re-evaluation in a subsequent command anyway
-	missingRefs := tools.NewStringSet()
-	for _, cachedRef := range cachedRemoteRefs {
-		found := false
-		for _, realRemoteRef := range actualRemoteRefs {
-			if cachedRef.Type == realRemoteRef.Type && cachedRef.Name == realRemoteRef.Name {
-				found = true
-				break
-			}
-		}
-		if !found {
-			missingRefs.Add(cachedRef.Name)
-		}
-	}
-
-	if len(missingRefs) > 0 {
-		// Use only the non-missing refs as 'from' points
-		commits := make([]string, 1, len(cachedRemoteRefs)+1)
-		commits[0] = refTo
-		for _, cachedRef := range cachedRemoteRefs {
-			if !missingRefs.Contains(cachedRef.Name) {
-				commits = append(commits, "^"+cachedRef.Sha)
-			}
-		}
-		return []string{"--stdin"}, commits
-	} else {
-		// Safe to use cached
-		return []string{refTo, "--not", "--remotes=" + remoteName}, nil
-	}
-}
-
-// revListShas uses git rev-list to return the list of object sha1s
-// for the given ref. If all is true, ref is ignored. It returns a
-// channel from which sha1 strings can be read.
-func revListShas(refLeft, refRight string, opt *ScanRefsOptions) (*StringChannelWrapper, error) {
-	refArgs := []string{"rev-list", "--objects"}
-	var stdin []string
-	switch opt.ScanMode {
-	case ScanRefsMode:
-		if opt.SkipDeletedBlobs {
-			refArgs = append(refArgs, "--no-walk")
-		} else {
-			refArgs = append(refArgs, "--do-walk")
-		}
-
-		refArgs = append(refArgs, refLeft)
-		if refRight != "" && !z40.MatchString(refRight) {
-			refArgs = append(refArgs, refRight)
-		}
-	case ScanAllMode:
-		refArgs = append(refArgs, "--all")
-	case ScanLeftToRemoteMode:
-		args, commits := revListArgsRefVsRemote(refLeft, opt.RemoteName)
-		refArgs = append(refArgs, args...)
-		if len(commits) > 0 {
-			stdin = commits
-		}
-	default:
-		return nil, errors.New("scanner: unknown scan type: " + strconv.Itoa(int(opt.ScanMode)))
-	}
-
-	// Use "--" at the end of the command to disambiguate arguments as refs,
-	// so Git doesn't complain about ambiguity if you happen to also have a
-	// file named "master".
-	refArgs = append(refArgs, "--")
-
-	cmd, err := startCommand("git", refArgs...)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(stdin) > 0 {
-		cmd.Stdin.Write([]byte(strings.Join(stdin, "\n")))
-	}
-
-	cmd.Stdin.Close()
-
-	revs := make(chan string, chanBufSize)
-	errchan := make(chan error, 5) // may be multiple errors
-
-	go func() {
-		scanner := bufio.NewScanner(cmd.Stdout)
-		for scanner.Scan() {
-			line := strings.TrimSpace(scanner.Text())
-			if len(line) < 40 {
-				continue
-			}
-
-			sha1 := line[0:40]
-			if len(line) > 40 {
-				opt.SetName(sha1, line[41:len(line)])
-			}
-			revs <- sha1
-		}
-
-		stderr, _ := ioutil.ReadAll(cmd.Stderr)
-		err := cmd.Wait()
-		if err != nil {
-			errchan <- fmt.Errorf("Error in git rev-list --objects: %v %v", err, string(stderr))
-		} else {
-			// Special case detection of ambiguous refs; lower level commands like
-			// git rev-list do not return non-zero exit codes in this case, just warn
-			ambiguousRegex := regexp.MustCompile(`warning: refname (.*) is ambiguous`)
-			if match := ambiguousRegex.FindStringSubmatch(string(stderr)); match != nil {
-				// Promote to fatal & exit
-				errchan <- fmt.Errorf("Error: ref %s is ambiguous", match[1])
-			}
-		}
-		close(revs)
-		close(errchan)
-	}()
-
-	return NewStringChannelWrapper(revs, errchan), nil
 }
 
 // revListIndex uses git diff-index to return the list of object sha1s
@@ -498,7 +328,7 @@ func revListIndex(atRef string, cache bool, indexMap *indexFileMap) (*StringChan
 // under the blobSizeCutoff will be ignored. revs is a channel over
 // which strings containing git sha1s will be sent. It returns a channel
 // from which sha1 strings can be read.
-func catFileBatchCheck(revs *StringChannelWrapper) (*StringChannelWrapper, error) {
+func catFileBatchCheck(revs <-chan string, errs <-chan error) (*StringChannelWrapper, error) {
 	cmd, err := startCommand("git", "cat-file", "--batch-check")
 	if err != nil {
 		return nil, err
@@ -545,11 +375,11 @@ func catFileBatchCheck(revs *StringChannelWrapper) (*StringChannelWrapper, error
 	}()
 
 	go func() {
-		for r := range revs.Results {
+		for r := range revs {
 			cmd.Stdin.Write([]byte(r + "\n"))
 		}
-		err := revs.Wait()
-		if err != nil {
+
+		if err := tools.CollectErrsFromChan(errs); err != nil {
 			// We can share errchan with other goroutine since that won't close it
 			// until we close the stdin below
 			errchan <- err
